Let the user choose the method to calculate
I let the user enter the number to determine what kind of method the user want to calculate, if the user enter the number is not in the range then let let the user enter again, and I also let the user enter the number of scale, and it will be used if the number is indivisible, then determine whether the scale number which the user enter is greater than zero, if the number is not greater than zero then let the user enter again.  Finally call the Calculator method which is in the PrecisionNumber class.